Change predict to accept functions
