+++ ftc/mfa/mfa.py
@@ -1,3 +1,5 @@
+from functools import reduce
+
 from ftc.mfa.polytope import Hypercube, Polytope
 
 
@@ -13,11 +15,8 @@ class MFA:
         vertices = ubox.vertices.map(self.create_distribute(t, state))
         return Polytope(vertices), nu[2:].ravel()
 
-    def predict(self, tspan, lmbd, scaling_factor=1.0):
-        ubox = self.ubox.map(lambda u_min, u_max: (lmbd * u_min, lmbd * u_max)).map(
-            lambda u_min, u_max: shrink(u_min, u_max, scaling_factor)
-        )
-
+    def predict(self, tspan, fns):
+        ubox = reduce(lambda ubox, fn: ubox.map(fn), (self.ubox, *fns))
         return self.is_success(map(lambda t: self.get_polynus(t, ubox), tspan))
 
     def create_distribute(self, t, state):
@@ -26,11 +25,3 @@ class MFA:
             return nu
 
         return distribute
-
-
-def shrink(u_min, u_max, scaling_factor=1.0):
-    mean = (u_min + u_max) / 2
-    width = (u_max - u_min) / 2
-    u_min = mean - scaling_factor * width
-    u_max = mean + scaling_factor * width
-    return u_min, u_max
